US_#_Tsk_#: Code clean, restructured and nodes updated
