+++ ui/src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import App from './App';
+import App from './Components/App';
 import {HashRouter, Route, Switch} from 'react-router-dom';
 
 import * as serviceWorker from './serviceWorker';
